refactor: implement BetManager for improved bet tracking and integrate into CheckAction and RaiseAction
