+++ pvm/ts/src/types/interfaces.ts
@@ -1,4 +1,17 @@
+import { Turn } from "../engine/types";
+
 export interface ICardDeck {
     shuffleDeck(): void;
     drawCard(): string;
+}
+
+export interface IBetManager {
+    add(action: Turn): void;
+    addTurns(turns: Turn[]): void;
+    current(): bigint;
+    delta(): bigint;
+    getTotalBetsForPlayer(playerId: string): bigint;
+    getBets(): Map<string, bigint>;
+    getLargestBet(): bigint;
+    getLastAggressor(start?: number): bigint;
 }
